requirements.txt
pytelegrambotapi
djangorestframework
python-decouple
psycopg2-binary
telebot-calendar

